+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 0.0.2
-- Changes supporting CloudHarvestCoreTasks 0.6.0
+- Changes supporting CloudHarvestCoreTasks 0.6.3
+- The `available_templates` information is now included in the heartbeat
+- Added `agent/list_plugins` and `agent/install_plugins` endpoints
+- Replaced `pstar` with `accounts` key in heartbeat
+- Added `harvest.templates` report
+- Updated some templates
 
 # 0.0.1
 - Initial implementation